changes are done with methods
